+++ assets/Ice.gcloud
@@ -1,0 +1,1 @@
+/home/viktor/OneDrive/PROJECTS/PERSONAL/GausClouds/Ice.gcloud
